// src/app/models/iusuario-sesion.ts
export interface IUsuarioSesion {
    empNd: number;
    empNdFil: number;
    eplNd: number;
    libConsecutivo: number;
    nombreUsuario: string;
    usuUsuario: string;
    primerApellidoCandidato: string;
    primerNombreCandidato: string;
    segundoApellidoCandidato: string;
    segundoNombreCandidato: string;
    tdcTd: string;
    tdcTdEpl: string;
    tdcTdFil: string;
}